Provide /var/lib/openqa/share mount point in helm charts
Some work on the worker charts. The worker container is configured to wait for
the webui pods before try to run itself. in its configuration will try to
connect using the internal Kubernetes network. However it fails to connect as
for now. This is might be due to the ingress-nginx and because the webui
container sets up an apache proxy.

poo: https://progress.opensuse.org/issues/183608
Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>