Demo openapi basic struct with a get and a post
The main component is the so-called `OpenAPI specification` which it can be
found in this PR under /api.json. This can be also a yaml for the record.
The next significant part is "x-mojo-to" which takes some route. Now because
the path for our API definitions is in lib/OpenQA/WebAPI/Controller/API/V1 I
have to add it in the `$self->routes->namespaces`. Otherwise the app couldnt
find the controller.

https://progress.opensuse.org/issues/186675
Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>