openapiv3: Demo of a simple API validation
The following finding are for GET methods. I assume that POST might be a
little more complicated.

Taking the previous commit and applying the following changes:
- switch to openapi v3 format
- convert json to yaml
- use validation via the yaml definition

The swagger still works. The yaml definition contains the `parameters` section
which makes the options available to the Swagger UI.
For a get method the current validation seems adequated.

The main differences from v2 to v3 was the yaml format:
- responses: use context instead of schema.
- servers: use an object to define it.
- start document and replace swagger with openapi

https://progress.opensuse.org/issues/186675
Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>